+++ html/webrtc.js
@@ -315,6 +315,8 @@ function getOffer() {
 }
 */
 
+// Stack Overflow より引用: https://stackoverflow.com/a/52760103
+// https://stackoverflow.com/questions/52738290/how-to-remove-video-codecs-in-webrtc-sdp
 function removeCodec(orgsdp, codec) {
   const internalFunc = (sdp) => {
     const codecre = new RegExp('(a=rtpmap:(\\d*) ' + codec + '\/90000\\r\\n)');
@@ -434,4 +436,8 @@ function moterStop() {
     moter4 : 0,
     moter5 : 0
   });
+}
+
+function play() {
+  remoteVideo.play();
 }